Split json_packets into status.py and data.py
