+++ modules/telemetry/telemetry.py
@@ -14,7 +14,10 @@ from pathlib import Path
 from signal import signal, SIGTERM
 from time import sleep
 from typing import Any, TypeAlias
-import modules.telemetry.json_packets as jsp
+from types import FrameType
+
+from modules.telemetry.data import TelemetryData
+from modules.telemetry.status import TelemetryStatus, MissionState, ReplayState
 import modules.telemetry.websocket_commands as wsc
 from modules.misc.config import Config
 from modules.telemetry.replay import TelemetryReplay
@@ -24,7 +27,6 @@ from modules.telemetry.utils import (
     ParsedTransmission,
 )
 from modules.telemetry.errors import MissionNotFoundError, AlreadyRecordingError, ReplayPlaybackError
-from types import FrameType
 
 # Types
 JSON: TypeAlias = dict[str, Any]
@@ -53,26 +55,28 @@ class Telemetry:
         version: str,
     ):
         super().__init__()
+        # Multiprocessing Queues to communicate with SerialManager and WebSocketHandler processes
         self.serial_status: Queue[str] = serial_status
         self.rn2483_radio_payloads: Queue[str] = rn2483_radio_payloads
         self.rn2483_radio_input: Queue[str] = rn2483_radio_input
         self.radio_signal_report: Queue[str] = radio_signal_report
         self.telemetry_json_output: Queue[JSON] = telemetry_json_output
         self.telemetry_ws_commands: Queue[list[str]] = telemetry_ws_commands
-        
+
         self.config = config
         self.version = version
 
+        # Telemetry Status holds the current status of the telemetry backend
         # Telemetry Data holds the last few copies of received data blocks stored under the subtype name as a key.
-        self.status: jsp.StatusData = jsp.StatusData()
-        self.telemetry_data: jsp.TelemetryData = jsp.TelemetryData(self.config.telemetry_buffer_size)
+        self.status: TelemetryStatus = TelemetryStatus()
+        self.telemetry_data: TelemetryData = TelemetryData(self.config.telemetry_buffer_size)
 
-        # Mission System
+        # Mission File System
         self.missions_dir = Path.cwd().joinpath("missions")
         self.missions_dir.mkdir(parents=True, exist_ok=True)
         self.mission_path: Path | None = None
 
-        # Mission Recording
+        # Mission Recording (not in use)
         self.mission_recording_file: BufferedWriter | None = None
         self.mission_recording_buffer: bytearray = bytearray(b"")
 
@@ -117,7 +121,7 @@ class Telemetry:
 
             # Switch data queues between replay and radio depending on mission state
             match self.status.mission.state:
-                case jsp.MissionState.RECORDED:
+                case MissionState.RECORDED:
                     while not self.replay_output.empty():
                         self.process_transmission(self.replay_output.get())
                         self.update_websocket()
@@ -139,7 +143,7 @@ class Telemetry:
 
     def reset_data(self) -> None:
         """Resets all live data on the telemetry backend to a default state."""
-        self.status = jsp.StatusData()
+        self.status = TelemetryStatus()
         self.telemetry_data.clear()
 
     def parse_serial_status(self, command: str, data: str) -> None:
@@ -150,15 +154,15 @@ class Telemetry:
             case "rn2483_connected":
                 self.status.rn2483_radio.connected = bool(data)
             case "rn2483_port":
-                if self.status.mission.state != jsp.MissionState.DNE:
+                if self.status.mission.state != MissionState.DNE:
                     self.reset_data()
                 self.status.rn2483_radio.connected_port = data
 
                 match self.status.rn2483_radio.connected_port:
                     case "":
-                        self.status.mission.state = jsp.MissionState.DNE
+                        self.status.mission.state = MissionState.DNE
                     case _:
-                        self.status.mission.state = jsp.MissionState.LIVE
+                        self.status.mission.state = MissionState.LIVE
             case _:
                 return None
 
@@ -221,13 +225,13 @@ class Telemetry:
         # Set replay status based on speed
         # If mission is not recorded, replay should be in DNE state.
         # if else, set to pause/playing based on speed
-        if self.status.mission.state != jsp.MissionState.RECORDED:
-            self.status.replay.state = jsp.ReplayState.DNE
+        if self.status.mission.state != MissionState.RECORDED:
+            self.status.replay.state = ReplayState.DNE
         elif speed == 0.0:
-            self.status.replay.state = jsp.ReplayState.PAUSED
+            self.status.replay.state = ReplayState.PAUSED
             self.replay_input.put(f"speed {speed}")
         else:
-            self.status.replay.state = jsp.ReplayState.PLAYING
+            self.status.replay.state = ReplayState.PLAYING
             self.replay_input.put(f"speed {speed}")
 
     def stop_replay(self) -> None:
@@ -258,7 +262,7 @@ class Telemetry:
         self.status.mission.name = mission_name
 
         # We are not to record when replaying missions
-        self.status.mission.state = jsp.MissionState.RECORDED
+        self.status.mission.state = MissionState.RECORDED
         self.status.mission.recording = False
 
         # Replay system
